# docs/api-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/products":{"get":{"operationId":"ProductsController_listProducts","parameters":[{"name":"page","required":false,"in":"query","description":"Page number (1-based)","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"minimum":1,"maximum":100,"default":10,"type":"number"}}],"responses":{"200":{"description":"Products retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedProductsResponseDto"}}}}},"summary":"Lists products with pagination","tags":["Products"]},"post":{"operationId":"ProductsController_createProduct","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"Product created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductResponseDto"}}}}},"summary":"Creates new product","tags":["Products"]}},"/api/v1/products/{id}":{"put":{"operationId":"ProductsController_updateProduct","parameters":[{"name":"id","required":true,"in":"path","description":"Unique resource identificator","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"Product updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductResponseDto"}}}},"404":{"description":"Product not found"}},"summary":"Updates existing product","tags":["Products"]},"get":{"operationId":"ProductsController_getProduct","parameters":[{"name":"id","required":true,"in":"path","description":"Unique resource identificator","schema":{"type":"string"}}],"responses":{"200":{"description":"Product retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductResponseDto"}}}},"404":{"description":"Product not found"}},"summary":"Gets product by ID","tags":["Products"]},"delete":{"operationId":"ProductsController_deleteProduct","parameters":[{"name":"id","required":true,"in":"path","description":"Unique resource identificator","schema":{"type":"string"}}],"responses":{"200":{"description":"Product deleted successfully"}},"summary":"Deletes product by ID","tags":["Products"]}},"/api/v1/products/{id}/reviews":{"get":{"operationId":"ReviewsController_listReviews","parameters":[{"name":"id","required":true,"in":"path","description":"Unique resource identificator","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","description":"Page number (1-based)","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"minimum":1,"maximum":100,"default":10,"type":"number"}},{"name":"sortBy","required":false,"in":"query","description":"Sort result by property","schema":{"default":"createdAt","type":"string","enum":["rating","createdAt"]}},{"name":"sortDirection","required":false,"in":"query","description":"Sort result in direction","schema":{"default":"DESC","type":"string","enum":["DESC","ASC"]}}],"responses":{"200":{"description":"Reviews retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedReviewsResponseDto"}}}}},"summary":"Lists reviews for a product with pagination","tags":["Reviews"]},"post":{"operationId":"ReviewsController_createReview","parameters":[{"name":"id","required":true,"in":"path","description":"Unique resource identificator","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":"Review created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewResponseDto"}}}}},"summary":"Creates new review for a product","tags":["Reviews"]}},"/api/v1/products/{id}/reviews/{reviewId}":{"put":{"operationId":"ReviewsController_updateReview","parameters":[{"name":"id","required":true,"in":"path","description":"Unique resource identificator","schema":{"type":"string"}},{"name":"reviewId","required":true,"in":"path","description":"Unique resource identificator","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateReviewDto"}}}},"responses":{"200":{"description":"Review updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewResponseDto"}}}},"404":{"description":"Review not found"}},"summary":"Updates existing review","tags":["Reviews"]},"delete":{"operationId":"ReviewsController_deleteReview","parameters":[{"name":"id","required":true,"in":"path","description":"Unique resource identificator","schema":{"type":"string"}},{"name":"reviewId","required":true,"in":"path","description":"Unique resource identificator","schema":{"type":"string"}}],"responses":{"200":{"description":"Review deleted successfully"}},"summary":"Deletes review by ID","tags":["Reviews"]}}},"info":{"title":"Product System","description":"API for manipulating products and reviews","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateProductResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier"},"name":{"type":"string","description":"Product name"},"price":{"type":"number","description":"Product price (in cents)"},"description":{"type":"string","description":"Product description"},"averageRating":{"type":"number","description":"Average rating if any reviews have been added to the product"}},"required":["id","name","price"]},"PaginatedProductsResponseDto":{"type":"object","properties":{"data":{"description":"List of products","type":"array","items":{"$ref":"#/components/schemas/CreateProductResponseDto"}},"total":{"type":"number","description":"Total number of products"},"page":{"type":"number","description":"Current page number"},"limit":{"type":"number","description":"Number of items per page"},"totalPages":{"type":"number","description":"Total number of pages"},"hasNext":{"type":"boolean","description":"Whether there is a next page"},"hasPrev":{"type":"boolean","description":"Whether there is a previous page"}},"required":["data","total","page","limit","totalPages","hasNext","hasPrev"]},"CreateProductDto":{"type":"object","properties":{"name":{"type":"string","description":"Product name"},"price":{"type":"number","description":"Product price (in cents)"},"description":{"type":"string","description":"Product description"}},"required":["name","price"]},"UpdateProductDto":{"type":"object","properties":{"name":{"type":"string","description":"Product name"},"price":{"type":"number","description":"Product price (in cents)"},"description":{"type":"string","description":"Product description"}}},"CreateReviewResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier"},"productId":{"type":"string","description":"Product ID"},"firstName":{"type":"string","description":"Reviewer first name"},"lastName":{"type":"string","description":"Reviewer last name"},"content":{"type":"string","description":"Review content"},"rating":{"type":"number","description":"Rating from 1 to 5"}},"required":["id","productId","firstName","lastName","content","rating"]},"PaginatedReviewsResponseDto":{"type":"object","properties":{"data":{"description":"List of reviews","type":"array","items":{"$ref":"#/components/schemas/CreateReviewResponseDto"}},"total":{"type":"number","description":"Total number of reviews"},"page":{"type":"number","description":"Current page number"},"limit":{"type":"number","description":"Number of items per page"},"totalPages":{"type":"number","description":"Total number of pages"},"hasNext":{"type":"boolean","description":"Whether there is a next page"},"hasPrev":{"type":"boolean","description":"Whether there is a previous page"}},"required":["data","total","page","limit","totalPages","hasNext","hasPrev"]},"CreateReviewDto":{"type":"object","properties":{"firstName":{"type":"string","description":"Reviewer first name"},"lastName":{"type":"string","description":"Reviewer last name"},"content":{"type":"string","description":"Review content"},"rating":{"type":"number","description":"Rating from 1 to 5","minimum":1,"maximum":5}},"required":["firstName","lastName","content","rating"]},"UpdateReviewDto":{"type":"object","properties":{"firstName":{"type":"string","description":"Reviewer first name"},"lastName":{"type":"string","description":"Reviewer last name"},"content":{"type":"string","description":"Review content"},"rating":{"type":"number","description":"Rating from 1 to 5","minimum":1,"maximum":5}}}}}}